[hvgl] Fix a warning

+++ src/hb-aat-var-hvgl-table.cc
@@ -81,7 +81,7 @@ PartShape::get_path_at (const struct hvgl &hvgl,
       float coord = _.second;
       bool pos = coord > 0;
       if (!coord) continue;
-      double scalar = fabs(coord);
+      double scalar = (double) fabs(coord);
       unsigned column_idx = axis_index * 2 + pos;
 
       const auto delta = deltas.get_column (column_idx, axisCount, segmentCount);
